requirements.txt
fastmcp
fastapi
uvicorn
pydantic
python-dotenv
google-genai
mcp
openai
